Graph plane parameter discovery
