+++ crates/hulkz/tests/common/mod.rs
@@ -1,0 +1,6 @@
+//! Shared test utilities for hulkz integration tests.
+
+/// Helper to create a unique namespace for test isolation.
+pub fn test_namespace(name: &str) -> String {
+    format!("test_{}_{}", name, std::process::id())
+}
